Moved valid classes into the module 🎅
