--- lib/seed_builder/valid_integer.rb
@@ -1,7 +1,0 @@
-class ValidInteger
-  def initialize model_object:, key:
-  end
-  def generate
-    rand(9999)
-  end
-end
